SortedSets:zRange
